More refactoring / using new styled button.
