+++ src/components/organisms/Quiz Components/NoQuizConfirm/index.js
@@ -1,0 +1,1 @@
+export { default } from "./NoQuizConfirm";
